separate load data and load batches functions
